+++ feature/links/application/hooks/queries/browser/useOpenBrowser.ts
