+++ feature/links/application/hooks/queries/link/useLinks.ts
@@ -1,7 +1,6 @@
+import { getLinksPreview } from "feature/links/application/services";
+import { type LinkPreview } from "feature/links/types/links";
 import useSWR from "swr";
-
-import { getLinksPreview } from "../service/linkServices";
-import { type LinkPreview } from "../types/links";
 
 type Purpose = "top-view" | "swipe";
