+++ feature/links/application/services/linkServices.ts
@@ -1,8 +1,8 @@
 import { type Session } from "@supabase/supabase-js";
 
 import supabase from "@/lib/supabase";
-import { parseUrl } from "../infrastructure/utils";
-import { type LinkPreview } from "../types/links";
+import { parseUrl } from "../../infrastructure/utils";
+import { type LinkPreview } from "../../types/links";
 
 export const getLinksPreview = async (
   limit: number = 5,
