+++ feature/links/components/cards/HorizontalCard.tsx
@@ -2,7 +2,7 @@ import { Image, View } from "react-native";
 
 import { PressableCard } from "@/components/pressable/PressableCard";
 import { ThemedText } from "@/components/text/ThemedText";
-import { useOpenBrowser } from "../../hooks/useOpenBrowser";
+import { useOpenBrowser } from "../../application/hooks/queries/browser/useOpenBrowser";
 import { type OGData } from "../../types/links";
 import { ErrorCard } from "./ErrorCard";
